bsd: add links for old sources
I found at TUHS an amazing list of old resources:
sources of old versions of BSD (and other systems).
Not only I need to document these links, but maybe it would be wise to
fetch the archives and store them in my personal cloud.  Then, I have to
start and stufy them!