kernel: import docs with hints installing LFS
"Linux From Scratch" is a rather useful and important project.
Normally, LFS is installed on a desktop or another server.
I found a developer who tried (with success) to install LFS on different
hypervisors, like XEN and EC2.  He included a few notes, which I import
here for easier access.

The documents was found at:

 - https://czep.net/17/lfs-on-ec2.html
 - https://czep.net/pub/ec2-hint.txt
 - https://czep.net/pub/xen-hint.txt