haskell: basic data types supplied by modules
Inspired by an article of Johannes Waldmann,
"How Do We Store Data, Then?"[1], I collected and documented all the
basic data types available in Haskell, to support algorithms that depend
on the non-efficient default type of `List`.  The article gives a nice
summary of which types are available and in which cases they are useful.

The situation is especially bad with `Strings`, and this probably needs
a special section of documentation.

[1]:	https://www.imn.htwk-leipzig.de/~waldmann/etc/untutorial/data/