ansic: list of nm(1) symbols
The output of nm(1) is rather terse and, sometimes, confusing.
Here is a quick reference of the meaning of these symbols.

Of course, I could use the manual page of nm(1), but easily and quickly
find this information is not so easy, as this summarised table.

The list is copied from the series of articles, "Driving Compilers".

[1]:    https://fabiensanglard.net/dc/